add removeNode Percentile accessors and create new Network to mutate
The old method of mutation would modify the original network instead of
creating a new network to then mutate based on the old network. 

The getter/setter for the removeNode functions were modified to allow
for the remove node to have it's own percentile. 

a method was created to change all of the percentiles at once because
the individual mutators can now have their own calculate number of nodes
to change as each mutation may need to do different amounts of work.